ot.lp reorganise to avoid def in __init__
